Split Sensors into standard Sensors and Sensors_advanced Only digital and analog sensors are in base sensors module, and DHT, USS and MPU are in advanced sensors.
